Prepare linux dependant code
