+++ netutil/netutil_linux.go
@@ -1,0 +1,3 @@
+// +build linux
+
+package netutil
